+++ src/src/main/java/sem3pl/dei/isep/ipp/pt/lapr3/application/menus/operationsUI/OperationRegisterUI.java
@@ -47,7 +47,7 @@ public class OperationRegisterUI {
             String format = "dd-MM-yyyy";
             Date date = Utils.readDate("Date", format);
 
-            int quantity = Utils.readInt("Quantity");
+            double quantity = Utils.readDouble("Quantity");
 
             List<Unidade> unidades = unidadeController.getUnidades();
             int unidade = Utils.selectUnidades("Choose the unity",unidades);
